Texture advanced but not done
Texture has a strange problem.
I can bind the texture, and it clearly creates an effect on the
rectangle viewport of the texture. But I can not seem to
obtain the contents of the image. I thought it is related to
size issues or some parameters of the texture. I am not
sure why