Refactored app window, written cube and events tutorials
This commit introduces several interesting changes.
1. Refactored the use of application window.
Now the skeleton window is delegated to window.py in
utils. When there is a need to change or go beyond
what's provided, simply subclass it like in
app.py of glevents.

2. Tutorial on drawing a cube is done

3. Tutorial on receiving events is done as well.

Next is the lightening. It should not be too difficult
to handle lightnening from now on.

Then I will give a break and deal with ptmviewer.