Lighting directions are all wrong but works
Some unknown reason directions are all wrong.
When the light object goes behind the object
it is the front side that is lit.
It is either shader or the code is at fault here.
But at least the object is lit.
The ambient light can be parametrized to use coefficients
as well for better control.
Diffuse light and the specular light can be used in
the same manner as well.