There is a problem with attenuation values
There is a problem with attenuation values they simply cancel out
the fragment value.
They either set it to white entirely. And it does not matter the
specular or diffuse color. They simply cancel it.
There is probably a problem with their correct setting.
The good news is that the shaders are not at fault here.